+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 1.0.1
+- Add optional `maxDecimals` to `formatterFromCurrency(Currency)` function
+- Add optional `maxDecimals` to `Currency.emptyPlaceholder` extension
+- Fix zero formatting in example
+- Update tests
 
+## 1.0.0
 - Initial version.